\encoding{UTF-8}
\name{FPTP2AV-package}
\alias{FPTP2AV}
\docType{package}
\title{Simulate Election Outcomes for AV Given FPTP Votes.
}
\description{
A program to simulate election outcomes under the alternative vote system using the actual FPTP votes.
}
\details{
\tabular{ll}{
Package: \tab FPTP2AV\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-12\cr
License: \tab LGPL (>= 2.0, < 3)\cr
}
}
\author{
Daniel Marcelino <dmarcelino@live.com>

}
\references{
Marcelino, Daniel (2013). \emph{SciencesPo: A Tool Set for Analyzing Political Behaviour Data}. Available at SSRN: \url{http://dx.doi.org/10.2139/ssrn.2320547}
}

\keyword{ package } 
\keyword{ fptp2av }
\seealso{
 \code{\link{fptp2av}}
}
\examples{
## Run the algorithm
 \dontrun{fptp2av(data, link=party_chains)}
}

