% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{on2011}
\alias{on2011}
\title{2011 Ontario Election Results}
\format{A data frame with 655 observations on the following 4 variables.}
\source{
\url{http://www.elections.on.ca/en-ca}
}
\usage{
data(on2011)
}
\description{
Official results of the 2011 Ontario election as available on Elections Ontario.

\itemize{
  \item Party names Name of the riding.
  \item riding Name of the riding.
  \item party Acronym version of the party names.
  \item votes Quantity of votes.}
}
\keyword{datasets}

