% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{readWriteCSV}
\alias{readWriteCSV}
\title{Read/Write Multiple csv Files at a Time}
\usage{
readWriteCSV(files, a.names = NULL, l.name = NULL, list = TRUE, pos = 1,
  envir = as.environment(pos))
}
\arguments{
\item{files}{csv file(s) to read.}

\item{a.names}{object names to assign the csv file(s) to.  If NULL assigns
the csv to the name(s) of the csv file(s) in the global environment.}

\item{l.name}{A character vector of names to assign to the csv files
(dataframes) being read in.  Default (NULL) uses the names of the files in
the directory without the file extension.}

\item{list}{A character vector of length one to name the list being read in.
Default is \code{"L1"}.}

\item{pos}{where to do the removal. By default, uses the current environment.}

\item{envir}{the environment to use.}

\item{\dots}{data.frame object(s) to write to a file}
}
\description{
\code{mcsv_w} - Read and assign multiple csv files at the same time.
}

